Title: 홀수만 출력하기, Time: 106ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240618/홀수만 출력하기/print-only-odd-numbers.kt
@@ -38,7 +38,7 @@ fun main() {
     // }
 
     repeat(a) {
-        val num = br.readLine()
+        val num = br.readLine().toInt()
 
         if(num % 2 == 1 && num % 3 == 0)
             sb.append("$num\n")
